+++ osu.Game/osu.Game.csproj
@@ -130,7 +130,8 @@
     <Compile Include="GameModes\Play\Taiko\TaikoHitRenderer.cs" />
     <Compile Include="GameModes\Play\Taiko\TaikoPlayfield.cs" />
     <Compile Include="GameModes\Edit\EditSongSelect.cs" />
-    <Compile Include="Graphics\UserInterface\ComboCounter.cs" />
+    <Compile Include="GameModes\Play\ComboCounter.cs" />
+    <Compile Include="GameModes\Play\ComboResultCounter.cs" />
     <Compile Include="Graphics\UserInterface\RollingCounter.cs" />
     <Compile Include="GameModes\Play\Taiko\TaikoComboCounter.cs" />
     <Compile Include="Input\GlobalHotkeys.cs" />
